% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashland.R
\name{ashland}
\alias{ashland}
\title{input for genetic simulations}
\format{heterogenous list object}
\description{
This object is itself a list of objects that are necessary to run a forward-time and coalescent simulation.  The entire list is the output object of holoSimCell::def_grid_pred(), and is an essential parameter for holoSim::getpophist.cells().
}
