% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browsePapers.R
\name{browsePapers}
\alias{browsePapers}
\title{Setup a page to browse papers}
\usage{
browsePapers()
}
\description{
run with no parameters, it sets up a page that links to literature supplied for the workshop
}
